Fix Kotlin Java records guide tests

+++ guides/micronaut-java-records/kotlin/src/test/kotlin/example/micronaut/BookControllerTest.kt
@@ -60,7 +60,7 @@ class BookControllerTest(
             about
         )
         val book = bookRepository.save(b)
-        assertEquals(booksInsertedInDbByMigration + 1, bookRepository.count())
+        assertEquals((booksInsertedInDbByMigration + 1).toLong(), bookRepository.count())
         val client = httpClient.toBlocking()
         val books = client.retrieve(
             HttpRequest.GET<Any>("/books"),
